integrate geo-validity-check into geo
Note: some tests are failing
