+++ geo/Cargo.toml
@@ -40,6 +40,8 @@ home = { version = "=0.5.9", optional = true }
 approx = ">= 0.4.0, < 0.6.0"
 criterion = { version = "0.4", features = ["html_reports"] }
 geo-test-fixtures = { path = "../geo-test-fixtures" }
+# REVIEW: make optional? remove? It's used for some conformance tests which are somewhat redundant with the JTS tests
+geos = { version = "9.1.1", features = ["geo"] }
 jts-test-runner = { path = "../jts-test-runner" }
 pretty_env_logger = "0.4"
 rand = "0.8.0"
